refactor(approval-gate): extract lifecycle.rs from lib.rs
Second step of breaking lib.rs into focused modules. This commit
lifts the persisted-record lifecycle helpers — pure Value-blob
constructors and transitions, no I/O, no iii-sdk dependency — into
their own module.

Moved (lib.rs -> lifecycle.rs):
- is_terminal_status()
- build_pending_record()
- transition_record() + transition_record_with_now()
- collect_timed_out_for_sweep()
- maybe_flip_timed_out()

All public items stay re-exported from the crate root.
Pure move, no behavior change. 141 approval-gate tests pass;
all 8 downstream worker crates pass.
